+++ NEWS.md
@@ -10,3 +10,14 @@ NEW FEATURES
 
 * Replaced distance measure for clustering of phylogenomic profiles: from
 vegan::vegdist to labdsv::dsvdis 
+
+# syntenet 0.99.7
+
+NEW FEATURES
+
+* Added functions fasta2AAStringSetlist and gff2GRangesList to help users
+read multiple FASTA and GFF/GTF files as list of AAStringSet and
+GRangesList, respectively.
+
+* Updated vignette to instruct users on how to load FASTA and GFF/GTF files
+to the R session.